[refactor] Change DIR
related to #19
